django/requirements.txt
Django==2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-widget-tweaks==1.4.5
Pillow==6.0.0
psycopg2==2.6.2
PyMySQL==0.9.3
python-decouple==3.1
pytz==2019.1
six==1.12.0
